divise get_cmd file
