+++ server/functions/#cmd_functions.c#
@@ -1,0 +1,9 @@
+/*
+** cmd_functions.c for  in /Volumes/SD/ETNA/slack/my_slack/server/functions
+** 
+** Made by DEBELLEIX Jérémy
+** Login   <debell_j@etna-alternance.net>
+** 
+** Started on  Mon Apr 24 18:31:49 2017 DEBELLEIX Jérémy
+** Last update Mon Apr 24 18:31:49 2017 DEBELLEIX Jérémy
+*/
