+++ server/functions/#my_whisp#
@@ -1,0 +1,9 @@
+  
+   my_whisp for  in /Volumes/SD/ETNA/slack/my_slack/server/functions
+   
+   Made by DEBELLEIX Jérémy
+   Login   <debell_e@etna-alternance.net>
+   
+   Started on  Mon Apr 24 18:43:30 2017 DEBELLEIX Jérémy
+   Last update Mon Apr 24 18:43:30 2017 DEBELLEIX Jérémy
+
